Add xrSession module
