+++ main.js
@@ -14,7 +14,6 @@ import {
   hoverCellEl,
   lastPickEl,
   btnMoveBoards,
-  aimInfoEl,
   turnEl,
   wireUI,
   setAimMode,
@@ -25,30 +24,35 @@ import {
   phase
 } from "./ui.js";
 
+import {
+  startAR,
+  getCellFromSelectEvent,
+  matrixFromTransform,
+  offsetLocalXZ,
+  xrSession,
+  getLastHitPose,
+  resetLastHitPose
+} from "./xrSession.js";
+
+export { startAR };
+
 const STORAGE_KEY = "ar-battleship-v1";
 
-let renderer, scene, camera;
-let xrSession = null;
-let localRefSpace = null;
-let viewerSpace = null;
-let hitTestSource = null;
-let reticle = null;
-let lastHitPose = null;
-
-let prevTime = null;
+export let renderer, scene, camera;
+export let reticle = null;
 
 // Zwei Boards
-let playerBoard = null;
-let enemyBoard = null;
+export let playerBoard = null;
+export let enemyBoard = null;
 
-let picker = null;
+export let picker = null;
 export let fleet = null;
 
 // Setup-State
 export let orientation = "H"; // "H" oder "V"
 
 // Runden-State
-let turn = "player"; // "player" | "ai"
+export let turn = "player"; // "player" | "ai"
 
 // --- NEU: KI-Zustand (Hunt/Target) ---
 let aiState = null;
@@ -60,9 +64,13 @@ let audioEnabled = true, hapticsEnabled = true;
 // Laden vorm AR-Start vormerken
 let pendingLoad = false;
 
-initGL();
-wireUI();
-diagnose().catch(()=>{});
+export function markPendingLoad() { pendingLoad = true; }
+export function checkPendingLoad() {
+  if (pendingLoad) {
+    pendingLoad = false;
+    setTimeout(() => loadState(), 200);
+  }
+}
 
 function initGL() {
   renderer = new THREE.WebGLRenderer({ canvas, antialias: true, alpha: true });
@@ -99,139 +107,12 @@ function onResize() {
   camera.updateProjectionMatrix();
 }
 
- 
-
-export async function startAR(mode = "regular") {
-  if (!navigator.xr) { statusEl.textContent = "WebXR nicht verfügbar. Bitte Meta/Quest-Browser verwenden."; await diagnose(); return; }
-  try {
-    const supported = await navigator.xr.isSessionSupported?.("immersive-ar");
-    if (supported === false) { statusEl.textContent = "Dieser Browser unterstützt 'immersive-ar' nicht. Bitte Quest-Browser updaten."; await diagnose(); return; }
-    const configs = mode === "safe"
-      ? [
-          { note: "SAFE: minimale Features", init: { requiredFeatures: [], optionalFeatures: [] } },
-          { note: "SAFE: optional hit-test", init: { requiredFeatures: [], optionalFeatures: ["hit-test"] } },
-        ]
-      : [
-          { note: "regular: hit-test + optional dom-overlay", init: { requiredFeatures: ["hit-test"], optionalFeatures: ["dom-overlay", "anchors", "hand-tracking"], domOverlay: { root: overlay } } },
-          { note: "regular-fallback: hit-test (kein dom-overlay)", init: { requiredFeatures: ["hit-test"], optionalFeatures: ["anchors", "hand-tracking"] } },
-        ];
-    let lastErr = null;
-    for (const cfg of configs) {
-      try { xrSession = await navigator.xr.requestSession("immersive-ar", cfg.init); statusEl.textContent = `AR gestartet (${cfg.note}).`; break; }
-      catch (e) { lastErr = e; }
-    }
-    if (!xrSession) throw lastErr || new Error("requestSession fehlgeschlagen (unbekannt)");
-
-    renderer.xr.setReferenceSpaceType("local");
-    await renderer.xr.setSession(xrSession);
-
-    xrSession.addEventListener("end", onSessionEnd);
-    xrSession.addEventListener("select", onSelect);
-    xrSession.addEventListener("squeezestart", onSqueeze);
-    xrSession.addEventListener("inputsourceschange", onInputSourcesChange);
-
-    localRefSpace = await xrSession.requestReferenceSpace("local");
-    viewerSpace = await xrSession.requestReferenceSpace("viewer");
-
-    try { hitTestSource = await xrSession.requestHitTestSource({ space: viewerSpace }); statusEl.textContent += " | hit-test aktiv."; }
-    catch { hitTestSource = null; statusEl.textContent += " | hit-test NICHT verfügbar."; }
-
-    btnStart.disabled = true; btnStartSafe.disabled = true;
-    btnReset.disabled = true;
-    setAimMode(aimMode);
-    setPhase("placement");
-    prevTime = null;
-    renderer.setAnimationLoop(onXRFrame);
-
-    // Sollte ein Laden vormerkt sein, jetzt ausführen
-    if (pendingLoad) { pendingLoad = false; setTimeout(() => loadState(), 200); }
-  } catch (err) {
-    statusEl.textContent = `AR-Start fehlgeschlagen: ${err?.name || "Error"} – ${err?.message || err}`;
-  }
-}
-
-function onSessionEnd() {
-  renderer.setAnimationLoop(null);
-  xrSession?.removeEventListener("select", onSelect);
-  xrSession?.removeEventListener("squeezestart", onSqueeze);
-  xrSession?.removeEventListener("inputsourceschange", onInputSourcesChange);
-  xrSession = null; hitTestSource = null; lastHitPose = null;
-  reticle.visible = false;
-  btnStart.disabled = false; btnStartSafe.disabled = false;
-  btnReset.disabled = !!playerBoard;
-  aimInfoEl.textContent = "";
-  setPhase("placement");
-}
-
-function onInputSourcesChange() {
-  if (!xrSession) return;
-  const src = pickActiveInputSource();
-  aimInfoEl.textContent = src
-    ? (aimMode === "controller" ? `Ray aktiv: ${src.handedness || "neutral"}` : "Zielen über Kopfblick.")
-    : (aimMode === "controller" ? "Kein Hand/Controller-Ray." : "Zielen über Kopfblick.");
-}
-
-function onXRFrame(time, frame) {
-  if (!frame) return;
-  lastXRFrame = frame;
-  // delta
-  if (prevTime == null) prevTime = time;
-  const dt = Math.min(0.1, (time - prevTime) / 1000);
-  prevTime = time;
-
-  if (phase === "placement") {
-    const results = hitTestSource ? frame.getHitTestResults(hitTestSource) : [];
-    if (results.length > 0) {
-      const pose = results[0].getPose(localRefSpace);
-      lastHitPose = pose && pose.transform;
-      if (lastHitPose) {
-        const m = new THREE.Matrix4().fromArray(lastHitPose.matrix ?? matrixFromTransform(lastHitPose));
-        reticle.visible = true; reticle.matrix.copy(m);
-      }
-    } else {
-      lastHitPose = null;
-      reticle.visible = false;
-    }
-  } else if (phase === "setup") {
-    picker.setBoard(playerBoard);
-    const cell = updateHover();
-    if (playerBoard && cell) {
-      const L = fleet.currentLength();
-      const valid = playerBoard.canPlaceShip(cell.row, cell.col, L, orientation);
-      playerBoard.showGhost(cell.row, cell.col, L, orientation, valid);
-    } else if (playerBoard) { playerBoard.clearGhost(); }
-  } else if (phase === "play") {
-    if (turn === "player") { picker.setBoard(enemyBoard); updateHover(); }
-    else { picker.setBoard(null); }
-  }
-
-  // FX updaten
-  playerBoard?.updateEffects?.(dt);
-  enemyBoard?.updateEffects?.(dt);
-
-  renderer.render(scene, camera);
-}
-
-function updateHover() {
-  if (!picker.board) return null;
-  if (aimMode === "gaze") {
-    const { changed, cell } = picker.updateFromCamera(camera);
-    if (changed) hoverCellEl.textContent = cell ? picker.board.cellLabel(cell.row, cell.col) : "–";
-    return picker.hoverCell || null;
-  } else {
-    const src = pickActiveInputSource();
-    if (!src) { picker.updateWithRay(new THREE.Vector3(1e6,1e6,1e6), new THREE.Vector3(0,-1,0)); return null; }
-    const pose = renderer.xr.getFrame().getPose(src.targetRaySpace, localRefSpace);
-    if (!pose) return null;
-    const { origin, dir } = originDirFromXRPose(pose);
-    const { changed, cell } = picker.updateWithRay(origin, dir);
-    if (changed) hoverCellEl.textContent = cell ? picker.board.cellLabel(cell.row, cell.col) : "–";
-    return picker.hoverCell || null;
-  }
-}
+initGL();
+wireUI();
+diagnose().catch(()=>{});
 
 /* ---------- Select: präziser Ray + Audio/Haptik/FX ---------- */
-function onSelect(e) {
+export function onSelect(e) {
   initAudio();
 
   if (phase === "placement") { placeBoardsFromReticle(); buzzFromEvent(e, 0.2, 30); playEarcon("placeBoard"); saveState(); return; }
@@ -313,15 +194,16 @@ function onSelect(e) {
   }
 }
 
-function onSqueeze() {
+export function onSqueeze() {
   if (phase === "setup") { rotateShip(); playEarcon("rotate"); saveState(); }
 }
 
 /* ---------- Boards platzieren ---------- */
 function placeBoardsFromReticle() {
-  if (!lastHitPose || playerBoard || enemyBoard) return;
+  const hitPose = getLastHitPose();
+  if (!hitPose || playerBoard || enemyBoard) return;
 
-  const baseM = new THREE.Matrix4().fromArray(lastHitPose.matrix ?? matrixFromTransform(lastHitPose));
+  const baseM = new THREE.Matrix4().fromArray(hitPose.matrix ?? matrixFromTransform(hitPose));
 
   playerBoard = new Board(0.50, 10, { baseColor: 0x0d1b2a, shipColor: 0x5dade2, showShips: true });
   playerBoard.placeAtMatrix(baseM);
@@ -383,9 +265,9 @@ export function moveBoards() {
   const tempPhase = phase;
   const tempTurn = turn;
   
-  playerBoard = null; enemyBoard = null;
-  fleet = null; aiState = null;
-  lastHitPose = null;
+    playerBoard = null; enemyBoard = null;
+    fleet = null; aiState = null;
+    resetLastHitPose();
   reticle.visible = true;
   btnReset.disabled = true;
   if (btnMoveBoards) btnMoveBoards.disabled = true;
@@ -579,62 +461,6 @@ export function resetAll() {
   if (btnMoveBoards) btnMoveBoards.disabled = true;
   statusEl.textContent = "Zurückgesetzt. Richte Reticle auf die Fläche und drücke Trigger zum Platzieren.";
   playEarcon("reset");
-}
-
-/* ---------- UI Helfer ---------- */
-/* ---------- XR Helpers ---------- */
-function pickActiveInputSource() {
-  if (!xrSession) return null;
-  let right = null, left = null, any = null;
-  for (const src of xrSession.inputSources) {
-    if (src.targetRayMode === "tracked-pointer") {
-      any = any || src;
-      if (src.handedness === "right") right = src;
-      else if (src.handedness === "left") left = src;
-    }
-  }
-  return right || left || any;
-}
-
-let lastXRFrame = null;
-
-function getCellFromSelectEvent(e, board) {
-  try {
-    if (!e || !board || !localRefSpace) return null;
-    const frame = e.frame || lastXRFrame;
-    if (!frame) return null;
-    if (!e.inputSource?.targetRaySpace) return null;
-    const pose = frame.getPose(e.inputSource.targetRaySpace, localRefSpace);
-    if (!pose) return null;
-    const m = new THREE.Matrix4().fromArray(pose.transform.matrix ?? matrixFromTransform(pose.transform));
-    const origin = new THREE.Vector3().setFromMatrixPosition(m);
-    const q = new THREE.Quaternion().setFromRotationMatrix(m);
-    const dir = new THREE.Vector3(0, 0, -1).applyQuaternion(q).normalize();
-    const hit = board.raycastCell(origin, dir);
-    return hit.hit ? { row: hit.row, col: hit.col } : null;
-  } catch { return null; }
-}
-
-function originDirFromXRPose(pose) {
-  const m = new THREE.Matrix4().fromArray(pose.transform.matrix ?? matrixFromTransform(pose.transform));
-  const origin = new THREE.Vector3().setFromMatrixPosition(m);
-  const q = new THREE.Quaternion().setFromRotationMatrix(m);
-  const dir = new THREE.Vector3(0, 0, -1).applyQuaternion(q).normalize();
-  return { origin, dir };
-}
-
-function matrixFromTransform(t) { return (new XRRigidTransform(t.position, t.orientation)).matrix; }
-
-function offsetLocalXZ(baseMatrix, dx, dz) {
-  const pos = new THREE.Vector3();
-  const quat = new THREE.Quaternion();
-  const scl = new THREE.Vector3();
-  new THREE.Matrix4().copy(baseMatrix).decompose(pos, quat, scl);
-  const offsetLocal = new THREE.Vector3(dx, 0, dz).applyQuaternion(quat);
-  pos.add(offsetLocal);
-  const out = new THREE.Matrix4();
-  out.compose(pos, quat, scl);
-  return out;
 }
 
 function allCells(n) { const arr = []; for (let r = 0; r < n; r++) for (let c = 0; c < n; c++) arr.push([r, c]); return arr; }
@@ -929,6 +755,6 @@ export function clearState() {
 }
 
 export function requestLoad() {
-  if (!xrSession) { pendingLoad = true; startAR("regular"); return; }
+  if (!xrSession) { markPendingLoad(); startAR("regular"); return; }
   loadState();
 }
